rag-workshop/requirements.txt
redis[hiredis]==5.0.1
openai==1.3.0
sentence-transformers==2.2.2
numpy==1.24.3
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.28.0